configured Button Component
