+++ src/react/components/Button.css
@@ -1,0 +1,18 @@
+.start-competition-button {
+    width: 27vh;
+    height: 6.5vh;
+    color: black;
+    font-size: 2vh;
+    border: none;
+    border-radius: 2vh;
+    margin: 1.5vh 0vh;
+}
+  
+.start-competition-button--primary {
+    background: #84afce;
+}
+  
+.start-competition-button--secondary {
+    background: red;
+}
+
